Se añaden validaciones y se pasa todo a float

+++ RandomVarGenerator/RandomVarGenerator/Generator.cs
@@ -8,19 +8,19 @@ namespace GeneradorDeNumerosAleatorios
 {
     public class Generator
     {
-        public decimal seed { get; set; }
-        public decimal a { get; set; }
-        public decimal c { get; set; }
-        public decimal M { get; set; }
-        public int type { get; set; }
+        public double seed { get; set; }
+        public double a { get; set; }
+        public double c { get; set; }
+        public double M { get; set; }
+        public double type { get; set; }
 
-        public List<decimal> Generate(int quantity)
+        public List<double> Generate(int quantity)
         {
-            List<decimal> result = new List<decimal>();
+            List<double> result = new List<double>();
 
             for (int i=0; i<quantity; i++)
             {
-                decimal rnd = ((this.a * this.seed) + this.c) % this.M;
+                double rnd = ((this.a * this.seed) + this.c) % this.M;
                 this.seed = rnd;
                 rnd = rnd / (this.M -1 );
                 rnd = Math.Truncate(rnd * 10000) / 10000;
@@ -30,9 +30,9 @@ namespace GeneradorDeNumerosAleatorios
             return result;
         }
 
-        public decimal NextRnd()
+        public double NextRnd()
         {
-            decimal rnd = ((this.a * this.seed) + this.c) % this.M;
+            double rnd = ((this.a * this.seed) + this.c) % this.M;
             this.seed = rnd;
             rnd = rnd / (this.M - 1);
             return rnd;
